CI: Move tests to test/unit
Change-type: patch
