+++ jest.config.js
@@ -2,7 +2,6 @@ module.exports = {
 	preset: "ts-jest",
 	testEnvironment: "node",
 	roots: [
-		"lib",
 		"test",
 	],
 };
