+++ lib/cursor.ts
@@ -1,9 +1,3 @@
-/*
- * Copyright (C) Balena.io - All Rights Reserved
- * Unauthorized copying of this file, via any medium is strictly prohibited.
- * Proprietary and confidential.
- */
-
 import type { Contract, JsonSchema, QueryOptions } from 'autumndb';
 import { cloneDeep } from 'lodash';
 import { v4 as uuid } from 'uuid';
