+++ package.json
@@ -36,7 +36,7 @@
     "lint": "balena-lint lib test && deplint",
     "lint:fix": "balena-lint --fix lib test",
     "test": "npm run lint && npm run test:unit",
-    "test:unit": "jest",
+    "test:unit": "jest test",
     "doc": "typedoc lib/ && touch docs/.nojekyll",
     "prepack": "npm run build"
   },
